pyproject.toml
[project]
name = "demo-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "groq>=0.31.0",
    "ollama>=0.5.3",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.4",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uvicorn>=0.35.0",
]
